requirements.txt
fastapi
uvicorn
python-multipart
google-generativeai
python-dotenv
scikit-learn
supabase
python-docx